Move displaying items to a separate component
Pass items to search results as a prop

Move show/hide insertion point

Move block insertion to menu

Fix block insertion
