+++ packages/block-editor/src/components/inserter/search-results.native.js
@@ -14,108 +14,26 @@ import {
  */
 import { useState, useEffect } from '@wordpress/element';
 import {
-	createBlock,
-	rawHandler,
-	store as blocksStore,
-} from '@wordpress/blocks';
-import { useSelect, useDispatch } from '@wordpress/data';
-import { withInstanceId } from '@wordpress/compose';
-import {
 	BottomSheet,
 	BottomSheetConsumer,
 	InserterButton,
-	getClipboard,
 } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import styles from './style.scss';
-import { store as blockEditorStore } from '../../store';
 
 const MIN_COL_NUM = 3;
 
-function InserterSearchResults( {
-	insertionIndex,
-	onDismiss,
-	onSelect,
-	isAppender,
-	clientId,
-	rootClientId,
-	shouldReplaceBlock,
-} ) {
+function InserterSearchResults( { items, onSelect, onClose } ) {
 	const [ numberOfColumns, setNumberOfColumns ] = useState( MIN_COL_NUM );
 	const [ itemWidth, setItemWidth ] = useState();
 	const [ maxWidth, setMaxWidth ] = useState();
 
-	const {
-		hideInsertionPoint,
-		insertDefaultBlock,
-		insertBlock,
-		clearSelectedBlock,
-		removeBlock,
-		showInsertionPoint,
-		resetBlocks,
-	} = useDispatch( blockEditorStore );
-
-	const {
-		items,
-		destinationRootClientId,
-		canInsertBlockType,
-		getBlockOrder,
-		getBlockCount,
-	} = useSelect( ( select ) => {
-		const selectBlockEditorStore = select( blockEditorStore );
-
-		const {
-			getInserterItems,
-			getBlockRootClientId,
-			getBlockSelectionEnd,
-		} = selectBlockEditorStore;
-
-		let targetRootClientId = rootClientId;
-		if ( ! targetRootClientId && ! clientId && ! isAppender ) {
-			const end = getBlockSelectionEnd();
-			if ( end ) {
-				targetRootClientId = getBlockRootClientId( end ) || undefined;
-			}
-		}
-
-		return {
-			items: getInserterItems( targetRootClientId ),
-			destinationRootClientId: targetRootClientId,
-			canInsertBlockType: selectBlockEditorStore.canInsertBlockType,
-			getBlockOrder: selectBlockEditorStore.getBlockOrder,
-			getBlockCount: selectBlockEditorStore.getBlockCount,
-		};
-	} );
-
-	const { getBlockType } = useSelect( ( select ) => select( blocksStore ) );
-
 	useEffect( () => {
-		if ( shouldReplaceBlock ) {
-			const count = getBlockCount();
-			// Check if there is a rootClientId because that means it is a nested replaceable block
-			// and we don't want to clear/reset all blocks.
-			if ( count === 1 && ! rootClientId ) {
-				// Removing the last block is not possilble with `removeBlock` action.
-				// It always inserts a default block if the last of the blocks have been removed.
-				clearSelectedBlock();
-				resetBlocks( [] );
-			} else {
-				const blockToReplace = getBlockOrder( destinationRootClientId )[
-					insertionIndex
-				];
-				removeBlock( blockToReplace, false );
-			}
-		}
-
-		showInsertionPoint( destinationRootClientId, insertionIndex );
-
 		Dimensions.addEventListener( 'change', onLayout );
-
 		return () => {
-			hideInsertionPoint();
 			Dimensions.removeEventListener( 'change', onLayout );
 		};
 	}, [] );
@@ -127,15 +45,6 @@ function InserterSearchResults( {
 		} = InserterButton.Styles.modalItem;
 		const { width } = InserterButton.Styles.modalIconWrapper;
 		return width + itemPaddingLeft + itemPaddingRight;
-	}
-
-	function onClose() {
-		// if should replace but didn't insert any block
-		// re-insert default block
-		if ( shouldReplaceBlock ) {
-			insertDefaultBlock( {}, destinationRootClientId, insertionIndex );
-		}
-		onDismiss();
 	}
 
 	function onLayout() {
@@ -163,54 +72,6 @@ function InserterSearchResults( {
 		}
 	}
 
-	/**
-	 * Processes the inserter items to check
-	 * if there's any copied block in the clipboard
-	 * to add it as an extra item
-	 */
-	function getItems() {
-		// Filter out reusable blocks (they will be added in another tab)
-		const itemsToDisplay = items.filter(
-			( { name } ) => name !== 'core/block'
-		);
-
-		const clipboardBlock = rawHandler( { HTML: getClipboard() } )[ 0 ];
-		const shouldAddClipboardBlock = canInsertBlockType(
-			clipboardBlock?.name,
-			destinationRootClientId
-		);
-
-		if ( shouldAddClipboardBlock ) {
-			const { name, icon } = getBlockType( clipboardBlock.name );
-			return [
-				{
-					name,
-					icon,
-					id: 'clipboard',
-					initialAttributes: clipboardBlock.attributes,
-					innerBlocks: clipboardBlock.innerBlocks,
-				},
-				...itemsToDisplay,
-			];
-		}
-
-		return itemsToDisplay;
-	}
-
-	function onBlockSelect( item ) {
-		const { name, initialAttributes, innerBlocks } = item;
-
-		const insertedBlock = createBlock(
-			name,
-			initialAttributes,
-			innerBlocks
-		);
-
-		insertBlock( insertedBlock, insertionIndex, destinationRootClientId );
-
-		onSelect();
-	}
-
 	return (
 		<BottomSheet
 			isVisible={ true }
@@ -226,7 +87,7 @@ function InserterSearchResults( {
 							key={ `InserterUI-${ numberOfColumns }` } //re-render when numberOfColumns changes
 							keyboardShouldPersistTaps="always"
 							numColumns={ numberOfColumns }
-							data={ getItems() }
+							data={ items }
 							ItemSeparatorComponent={ () => (
 								<TouchableWithoutFeedback accessible={ false }>
 									<View style={ styles.rowSeparator } />
@@ -239,8 +100,8 @@ function InserterSearchResults( {
 										item,
 										itemWidth,
 										maxWidth,
+										onSelect,
 									} }
-									onSelect={ onBlockSelect }
 								/>
 							) }
 							{ ...listProps }
@@ -260,4 +121,4 @@ function InserterSearchResults( {
 	);
 }
 
-export default withInstanceId( InserterSearchResults );
+export default InserterSearchResults;
